+++ src/StackerMedian.cxx
@@ -105,7 +105,7 @@ void StackerMedian::add_photo_to_stack(unsigned int file_index, int y_min, int y
         }
     }
     else    {
-        unsigned int value;
+        short int value;
         char color;
         for (int y = y_min; y < y_max; y++)  {
             for (int x = 0; x < m_width; x++)   {
